Fixed TVI and TVI24 streams (#34)
* tvi24: updated dynamic stream into a working one

We need to actually fetch the link each time, instead of harcoding
it, since there is an auth token.

* fix tvi stream

We need an auth token for the playlist now. While we probably should get
getting it our of the TVI player, we're going directly to the TVI24 player
instead, which has the token more readilly available (and the same token works
across channels, for now at least).